+++ src/interpreter/handlers/arithmetic.cc
@@ -1,9 +1,11 @@
+#include <cstdint>
+
 #include "interpreter/handlers/handlers.hh"
 
 bool interpreter::handlers::add_handler(cpu::BaseCPU& cpu)
 {
-  long long value2 = cpu.stack_pop();
-  long long value1 = cpu.stack_pop();
+  int64_t value2 = cpu.stack_pop();
+  int64_t value1 = cpu.stack_pop();
 
   cpu.stack_push(value1 + value2);
 
@@ -12,8 +14,8 @@ bool interpreter::handlers::add_handler(cpu::BaseCPU& cpu)
 
 bool interpreter::handlers::sub_handler(cpu::BaseCPU& cpu)
 {
-  long long value2 = cpu.stack_pop();
-  long long value1 = cpu.stack_pop();
+  int64_t value2 = cpu.stack_pop();
+  int64_t value1 = cpu.stack_pop();
 
   cpu.stack_push(value1 - value2);
 
@@ -22,8 +24,8 @@ bool interpreter::handlers::sub_handler(cpu::BaseCPU& cpu)
 
 bool interpreter::handlers::mul_handler(cpu::BaseCPU& cpu)
 {
-  long long value2 = cpu.stack_pop();
-  long long value1 = cpu.stack_pop();
+  int64_t value2 = cpu.stack_pop();
+  int64_t value1 = cpu.stack_pop();
 
   cpu.stack_push(value1 * value2);
 
@@ -32,8 +34,8 @@ bool interpreter::handlers::mul_handler(cpu::BaseCPU& cpu)
 
 bool interpreter::handlers::div_handler(cpu::BaseCPU& cpu)
 {
-  long long value2 = cpu.stack_pop();
-  long long value1 = cpu.stack_pop();
+  int64_t value2 = cpu.stack_pop();
+  int64_t value1 = cpu.stack_pop();
 
   cpu.stack_push(value1 / value2);
 
@@ -42,8 +44,8 @@ bool interpreter::handlers::div_handler(cpu::BaseCPU& cpu)
 
 bool interpreter::handlers::mod_handler(cpu::BaseCPU& cpu)
 {
-  long long value2 = cpu.stack_pop();
-  long long value1 = cpu.stack_pop();
+  int64_t value2 = cpu.stack_pop();
+  int64_t value1 = cpu.stack_pop();
 
   cpu.stack_push(value1 % value2);
